+++ src/main/java/com/team03/monew/controller/NotificationController.java
@@ -1,7 +1,6 @@
 package com.team03.monew.controller;
 
-import com.team03.monew.dto.CursorPageResponseNotificationDto;
-import com.team03.monew.dto.NotificationDto;
+import com.team03.monew.dto.junhyeok.CursorPageResponseNotificationDto;
 import com.team03.monew.service.NotificationService;
 import java.time.LocalDateTime;
 import java.util.UUID;
